libs/interview_eval/pyproject.toml
[project]
name = "interview-eval"
version = "0.0.1"
description = ""
readme = "README.md"
authors = [
    { name = "Juyoung Suk", email = "juyoung@kaist.ac.kr" },
]
requires-python = ">=3.12"
dependencies = [
    "black>=24.10.0",
    "click>=8.1.7",
    "isort>=5.13.2",
    "openai>=1.55.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
]

[tool.uv]
package = true

[project.scripts]
interview-eval = "interview_eval.cli:main"

[dependency-groups]
dev = [
    "black>=24.10.0",
    "isort>=5.13.2",
    "pytest>=8.3.3",
]

[tool.setuptools]
packages = ["interview_eval"]

